You have to explicitly set mass and add forces.
There is no longer a way to set mass and acceleration at creation time, as it kind of didn't make sense. You have to
explicitly add a mass, and add accelerations by applying forces (not implemented yet). Also thinking of doing the same
for velocity, so it can only be done by applying forces, which will change the acceleration, which in turn will
change the velocity. But right now I won't bother doing it.