math: Change vector2 to be a simple struct with standalone functions.
A vector is, theoretically, just a data structure. This changes the
implementation of vector2 to be a simple data structure, with only
constructors, copy operator, index operator and destructor as member
functions. All other operations now stands as free standing functions.

When this is done on all vector classes, which will happen soon, we can
easily change the vector to one structure, and have special functions
when needed for a vector (Like cross for a 3D vector).

Signed-off-by: Martin Ertsaas <martiert@gmail.com>